Fixed case when ajax or whatever update dom fragments with the same ids as before
If in the updated fragment an id was used as target by crapLoader before
the request, then the new element with the same id used as target too
would not be updated by crapLoader, because the first element with this id still
existed in `elementCached`.
So we need to flush this cache when crapLoader enters in "release" mode.